blog/requirements.txt
asgiref==3.5.1
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==12.0.1
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.2.0
mccabe==0.6.1
mixer==7.2.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
zipp==3.8.0
